requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
PyPDF2==3.0.1
python-dotenv==1.0.1
openai==1.60.2